requirements.txt

Flask==2.3.2
cohere
requests
PyJWT
